Ensure that subprocess.Popen calls happen in the activted virtualenv.
Without this the `awsebcli` python module was getting installed to
the location pointed out by `sys.executable` which in this scenario,
wherein the virtual env is activated via an exec/execfile of the
`activate_this.py` code, is not set to that of the virtualenv python
executable and hence the base system's `pip` kicks in and installs
it in the base system's python site-packages path. This further
causes an issue when the eb executable wrapper tries to reach it
at the virtualenv's `bin_location/eb` and results in an `eb` not
found error.